Printer bugfixes and carry a context during printing
Changes:
 - Fix issue of printing anonymous tuple constructors, e.g. "_#Make3(x,y,z)" in DafnyToYIL.
 - Add a `MethodType` type in YIL for:
  * Add support for diffing methods, function methods, etc. for correct printing of projects that don't separate spec / impl.
  * distinguish between static / nonstatic members and methods, function methods, predicates, etc.
 - Method bodies support a different set of syntax compared to functions. E.g. function returns are just expressions, but method
returns are return statements using the return keyword. In addition, function if-statements have a "then" keyword, but methods do not.
To distinguish between what set of syntax to print in Printer, we need contextual information of what kind of YIL method we're in (i.e.
whether it maps back to a Dafny function or method or predicate or lemma, ...). However, the printer does not carry a context by default.
If we trivially add a YIL.Context argument to every mutually recursive method in Printer, there's immediately a problem: YIL.Context is
the function of a YIL.Program, but the differ also uses the YIL Printer, and we lose access to a YIL.Program on the differ since it
operates on the diff of two programs. We fix this issue by adding a "PrintingContext" interface, which describes all contextual information
the Printer needs, in YIL.fs, and have the original Context implement a "PrintingContext". The differ can then pass in an "EmptyPrintingContext",
which is a trivial implementation of the printing context.