Major updates from the last few weeks
These are almost entirely untested and certainly still buggy.
But I'm having difficulties testing because
* my old test files have not been integrated into the new test harness yet
* YuccaModelDafny does not build for me yet since DafnyLiteLib was introduced

Undoing a few Bias-for-Action changes that Ruijie made
* remove TApplyPrimitive
* remove prefixing function/lemma names

Removed special treatment for types defined in DafnyLiteLib
That code was copied over from YIL in the YuccaCompiler, but it has turned out to be counter-productive for typeCart.
Some left-overs may still have to be removed.

Added feature for generating code for the functional approach
The structure of the code is very similar for both approaches. So I did both in the same class:
* Translation.Translator now takes a boolean flag for toggling between functional/relational
* A few key places in the translation use the flag.